Fixing routes

+++ app/api/patients/[id]/route.ts
@@ -1,24 +1,28 @@
-import { NextResponse } from "next/server";
+import { NextRequest, NextResponse } from "next/server";
 import { prisma } from "@/lib/prisma";
 
 // Get a specific patient
 export async function GET(
-  request: Request,
-  { params }: { params: { id: string } }
+  request: NextRequest
 ) {
   try {
+    const { searchParams } = new URL(request.url);
+    const id = searchParams.get("id");
+    if (!id) {
+      return NextResponse.json({ error: "ID parameter is required." }, { status: 400 });
+    }
     const patient = await prisma.patient.findUnique({
-      where: { id: params.id },
+      where: { id },
       include: {
         dietCharts: true,
         mealBoxes: true,
       },
     });
-
+    
     if (!patient) {
       return NextResponse.json({ error: "Patient not found" }, { status: 404 });
     }
-
+    
     return NextResponse.json(patient);
   } catch (error) {
     console.error(error);
@@ -32,12 +36,16 @@ export async function GET(
 // Update a patient
 export async function PUT(
   request: Request,
-  { params }: { params: { id: string } }
 ) {
   try {
+    const { searchParams } = new URL(request.url);
+    const id = searchParams.get("id");
+    if (!id) {
+      return NextResponse.json({ error: "ID parameter is required." }, { status: 400 });
+    }
     const json = await request.json();
     const patient = await prisma.patient.update({
-      where: { id: params.id },
+      where: { id },
       data: json,
     });
     return NextResponse.json(patient);
@@ -53,11 +61,16 @@ export async function PUT(
 // Delete a patient
 export async function DELETE(
   request: Request,
-  { params }: { params: { id: string } }
 ) {
   try {
+    const { searchParams } = new URL(request.url);
+    const id = searchParams.get("id");
+    if (!id) {
+      return NextResponse.json({ error: "ID parameter is required." }, { status: 400 });
+    }
+
     await prisma.patient.delete({
-      where: { id: params.id },
+      where: { id },
     });
     return NextResponse.json({ message: "Patient deleted successfully" });
   } catch (error) {
